+++ tests/DependencyInjection/SentryExtensionTest.php
@@ -29,6 +29,7 @@ use Sentry\Serializer\Serializer;
 use Sentry\Transport\TransportFactoryInterface;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Console\ConsoleEvents;
+use Symfony\Component\Debug\Exception\FatalErrorException;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\ParameterBag\EnvPlaceholderParameterBag;
@@ -185,7 +186,14 @@ abstract class SentryExtensionTest extends TestCase
         $optionsDefinition = $container->getDefinition('sentry.client.options');
         $expectedOptions = [
             'integrations' => [
-                new Definition(IgnoreErrorsIntegration::class, [['ignore_exceptions' => [FatalError::class]]]),
+                new Definition(IgnoreErrorsIntegration::class, [
+                    [
+                        'ignore_exceptions' => [
+                            FatalError::class,
+                            FatalErrorException::class,
+                        ],
+                    ],
+                ]),
                 new Reference('App\\Sentry\\Integration\\FooIntegration'),
             ],
             'default_integrations' => false,